Mess around with the pad colors.
These are now looking less bright and garish, IMHO (mk2 only). Note that
on the mk1 controller, which has no proper RGB, these will all be mapped
to just red, green and yellow; it's hard to improve on this without
messing up the color scheme for the mk2.

You can edit the colors yourself in the banks subpatch. This has a list
of eight values (4 for the 4 banks of pads, 4 for the 4 banks of
sequence launchers, in that order). Note that these are velocity values;
the mapping of these to actual RGB values is described in the "APC mini
mk2 - Communication Protocol" document available from the AKAI website.
You can also check the colors using the apcmini-help patch included in
the repository (see the color-animation subpatch).