Refactor pad color mapping, to make it easier to edit the colors.
The color scheme can now be adjusted more easily by editing the list of
8 velocities (4 for each of the pad banks, 4 for each of the sequence
banks) in the lower-right corner of the banks subpatch. There are also
some bang controls to generate random color schemes, if you're not
content with the defaults, or to return to the default color scheme
after you had enough of the random colorings.

Note that the random color generator also writes the generated scheme to
a list message below, from where you can copy it to the defaults color
list, which will make it the new default after saving the patch. In the
future we may want to add a proper color editor which saves the color
scheme in an external file, but this will do for now.