Improved MIDI mapping.
Remapped the pads and sequence triggers to channel 10-13 and all other
controls to channel 16. This makes it easier to use the patch if Koala
runs on the same system, which is complicated because Koala currently
reads MIDI data from all available MIDI inputs; check the new "Bugs and
Quirks" section in the documentation for details.