Refactoring.
Put all the control logic into the controls subpatch which also offers
easier access to settings that the user may want to modify, like the
grid colors.

MIDI channels for the grid and the other controls (channel 10-13 and 16
by default) are now configurable there as well. But note that if you
change these then you'll also have to edit the midiMapping.json file
accordingly.